Remove hardcoded styles for tables
Change hardcoded styles for tables row to CSS classes.

Fix notification message for multiple files upload success.

+++ admin/backups.php
@@ -9,6 +9,8 @@
  * @link http://get-simple.info/docs/restore-page-backup
  */
 
+declare(strict_types=1);
+
 // Setup inclusions
 $load['plugin'] = true;
 
@@ -17,63 +19,58 @@ include('inc/common.php');
 
 // Variable settings
 login_cookie_check();
-$path = GSBACKUPSPATH.'pages/';
-$counter = '0';
+$path = GSBACKUPSPATH . 'pages/';
+$counter = 0;
 $table = '';
 
 // delete all backup files if the ?deleteall session parameter is set
-if (isset($_GET['deleteall'])){
+if (isset($_GET['deleteall'])) {
 	// check for csrf
-	if (!defined('GSNOCSRF') || (GSNOCSRF == FALSE) ) {
-		$nonce = $_GET['nonce'];
-		if(!check_nonce($nonce, "deleteall")) {
-			die("CSRF detected!");
-		}
+	if (!defined('GSNOCSRF') || (GSNOCSRF == false)) {
+		$nonce = (string)filter_input(INPUT_GET, 'nonce');
+		if (!check_nonce($nonce, 'deleteall')) die('CSRF detected!');
 	}
 	$filenames = getFiles($path);
-	
+
 	foreach ($filenames as $file) {
-		if (file_exists($path . $file) ) {
-			if (isFile($file, $path, 'bak')) {
-				unlink($path . $file);
-			}
+		if (file_exists($path . $file)) {
+			if (isFile($file, $path, 'bak')) unlink($path . $file);
 		}
 	}
-	
+
 	$success = i18n_r('ER_FILE_DEL_SUC');
 }
 
 
 //display all page backups
 $filenames = getFiles($path);
-$count="0";
+$count = 0;
 $pagesArray_tmp = array();
-$pagesSorted=array(); 
+$pagesSorted = array(); 
 
-if (count($filenames) != 0) { 
+if (count($filenames) != 0) {
 	foreach ($filenames as $file) {
 		if (isFile($file, $path, 'bak')) {
 			$data = getXML($path .$file);
-			$status = $data->menuStatus;
-			$pagesArray_tmp[$count]['title'] = html_entity_decode($data->title, ENT_QUOTES, 'UTF-8');
-			$pagesArray_tmp[$count]['url'] = $data->url;
-			$pagesArray_tmp[$count]['date'] = $data->pubDate;
+			$pagesArray_tmp[$count]['title'] = html_entity_decode((string)$data->title, ENT_QUOTES, 'UTF-8');
+			$pagesArray_tmp[$count]['url'] = (string)$data->url;
+			$pagesArray_tmp[$count]['date'] = (string)$data->pubDate;
 			$count++;
 		}
 	}
-	$pagesSorted = subval_sort($pagesArray_tmp,'title');
+	$pagesSorted = subval_sort($pagesArray_tmp, 'title');
 }
 
 if (count($pagesSorted) != 0) {
 	foreach ($pagesSorted as $page) {
 		$counter++;
-		$table .= '<tr id="tr-'.$page['url'] .'" >';
+		$table .= '<tr id="tr-' . $page['url'] . '" >';
 
-		if ($page['title'] == '' ) $page['title'] = '[No Title] &nbsp;&raquo;&nbsp; <em>'. $page['url'] .'</em>';
+		if ($page['title'] == '') $page['title'] = '[No Title] &nbsp;&raquo;&nbsp; <em>' . $page['url'] . '</em>';
 
-		$table .= '<td class="pagetitle"><a title="'.i18n_r('VIEWPAGE_TITLE').' '. var_out($page['title']) .'" href="backup-edit.php?p=view&amp;id='. $page['url'] .'">'. cl($page['title']) .'</a></td>';
-		$table .= '<td style="width:80px;text-align:right;" ><span>'. shtDate($page['date']) .'</span></td>';
-		$table .= '<td class="delete" ><a class="delconfirm" title="'.i18n_r('DELETEPAGE_TITLE').' '. var_out($page['title']) .'?" href="backup-edit.php?p=delete&amp;id='. $page['url'] .'&amp;nonce='.get_nonce("delete", "backup-edit.php").'">&times;</a></td>';
+		$table .= '<td class="pagetitle"><a title="' . i18n_r('VIEWPAGE_TITLE') . ' ' . var_out($page['title']) . '" href="backup-edit.php?p=view&amp;id=' . $page['url'] . '">' . cl($page['title']) . '</a></td>';
+		$table .= '<td class="date"><span>' . shtDate($page['date']) . '</span></td>';
+		$table .= '<td class="delete"><a class="delconfirm" title="' . i18n_r('DELETEPAGE_TITLE') . ' ' . var_out($page['title']) . '?" href="backup-edit.php?p=delete&amp;id=' . $page['url'] . '&amp;nonce=' . get_nonce('delete', 'backup-edit.php') . '">&times;</a></td>';
 		$table .= '</tr>';
 	}
 }
@@ -88,17 +85,17 @@ get_template('header', cl($SITENAME) . ' &raquo; ' . i18n_r('BAK_MANAGEMENT'));
 
 	<div id="maincontent">
 		<div class="main">
-			<h3><?php i18n('PAGE_BACKUPS');?></h3>
+			<h3><?php i18n('PAGE_BACKUPS'); ?></h3>
 			<?php if ($counter > 0) { ?>
-				<div class="edit-nav"><a href="#" id="filtertable" accesskey="<?php echo find_accesskey(i18n_r('FILTER'));?>" ><?php i18n('FILTER'); ?></a> <a href="backups.php?deleteall&amp;nonce=<?php echo get_nonce("deleteall"); ?>" title="<?php i18n('DELETE_ALL_BAK');?>" accesskey="<?php echo find_accesskey(i18n_r('ASK_DELETE_ALL'));?>" class="confirmation"><?php i18n('ASK_DELETE_ALL');?></a></div>
+				<div class="edit-nav"><a href="#" id="filtertable" accesskey="<?php echo find_accesskey(i18n_r('FILTER'));?>" ><?php i18n('FILTER'); ?></a> <a href="backups.php?deleteall&amp;nonce=<?php echo get_nonce('deleteall'); ?>" title="<?php i18n('DELETE_ALL_BAK');?>" accesskey="<?php echo find_accesskey(i18n_r('ASK_DELETE_ALL'));?>" class="confirmation"><?php i18n('ASK_DELETE_ALL'); ?></a></div>
 				<div id="filter-search">
-					<form><input type="text" autocomplete="off" class="text" id="q" placeholder="<?php echo strip_tags(lowercase(i18n_r('FILTER'))); ?>..." /> &nbsp; <a href="pages.php" class="cancel"><?php i18n('CANCEL'); ?></a></form>
+					<form><input type="text" autocomplete="off" class="text" id="q" placeholder="<?php echo strip_tags(lowercase(i18n_r('FILTER'))); ?>..."> &nbsp; <a href="pages.php" class="cancel"><?php i18n('CANCEL'); ?></a></form>
 				</div>
 				<table id="editpages" class="highlight paginate">
-					<tr><th><?php i18n('PAGE_TITLE'); ?></th><th style="text-align:right;" ><?php i18n('DATE'); ?></th><th></th></tr>
+					<tr><th><?php i18n('PAGE_TITLE'); ?></th><th class="date"><?php i18n('DATE'); ?></th><th></th></tr>
 					<?php echo $table; ?>
 				</table>
-			<?php  } else { ?>
+			<?php } else { ?>
 				<div class="clearfix" style="height:40px;"></div>
 			<?php } ?>
